refactor: notification eventhandler
